+++ Game/Test/TestServerSideProtocol.cpp
@@ -84,7 +84,7 @@ namespace Test
 			uint8_t inputs[4] = { 1, 20, 77, 78 };
 			//TODO: Aggregate initialization is convenient, but automatically filling these default params would be cool, too.
 			//Add a packet factory, maybe?
-			net::ClientInputPacket packet{ def::protocol_id, 0, net::client_input, 1337, 2, util::countof(inputs), inputs };
+			net::ClientInputPacket packet{ def::protocol_id, 0, net::client_input, 1337, 2, static_cast<uint16_t>(util::countof(inputs)), inputs };
 			packet.IO<net::Write>(client_socket.send_buffer);
 		}
